requirements.txt
Django==1.6.2
South==0.8.4
argparse==1.2.1
django-jqgrid==dev
unittest-data-provider==1.0.1
wsgiref==0.1.2
